fixed stupid gcc warning

+++ firmware/at91sam7/core/openbeacon/debug_printf.c
@@ -27,7 +27,7 @@ simple_strtoul (const char *cp, char **endp, unsigned int base)
   if (*cp == '0')
     {
       cp++;
-      if ((*cp == 'x') && isxdigit (cp[1]))
+      if ((*cp == 'x') && isxdigit ((int)cp[1]))
 	{
 	  base = 16;
 	  cp++;
@@ -41,9 +41,9 @@ simple_strtoul (const char *cp, char **endp, unsigned int base)
     {
       base = 10;
     }
-  while (isxdigit (*cp) && (value = isdigit (*cp) ? *cp - '0' : (islower (*cp)
+  while (isxdigit ((int)*cp) && (value = isdigit ((int)*cp) ? *cp - '0' : (islower ((int)*cp)
 								 ?
-								 toupper (*cp)
+								 toupper ((int)*cp)
 								 : *cp) -
 			    'A' + 10) < base)
     {
